refactor: extract logger into internal
Also move the action-related logger into action-ln.
